+++ satellite/src/main/java/com/linbit/linstor/layer/drbd/resfiles/package-info.java
@@ -1,0 +1,4 @@
+@NonNullByDefault
+package com.linbit.linstor.layer.drbd.resfiles;
+
+import com.linbit.linstor.annotation.NonNullByDefault;