+++ server/src/main/java/com/linbit/linstor/storage/interfaces/layers/drbd/DrbdRscObject.java
@@ -48,7 +48,8 @@ public interface DrbdRscObject<RSC extends AbsResource<RSC>>
         DISK_REMOVING(1L << 6),
         INITIALIZED(1L << 7),
         FROM_BACKUP(1L << 8),
-        FORCE_NEW_METADATA(1L << 9);
+        FORCE_NEW_METADATA(1L << 9),
+        INVALIDATE(1L << 10);
 
         public final long flagValue;