+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Attempt to fix race-condition "Received unknown storage resource from satellite"
 - Prometheus: fix using peerId as peer address in api call duration statistics (GitHub #466)
 - H2 db should no longer prematurely shutdown
+- DRBD: recreate ext-metadata after backing disk needed to be re-created
 
 ## [1.32.3] - 2025-10-13